+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Update the Splunk On-Call integration to not retry HTTP 404 responses
 - Update the Slack integration to not retry when Slack returns 400 "invalid_token"
 - Improve Pushover notifications (include tags, period, last ping type etc.)
+- Update the Pushover integration to not retry when Pushover reports invalid user
 
 ### Bug Fixes
 - Fix "Ping Details" dialog to handle email bodies not yet uploaded to object storage